third_party/WebKit/LayoutTests/fast/dom/MutationObserver/observe-exceptions-expected.txt

Test that WebKitMutationObserver.observe throws exceptions appropriately

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS observer.observe() threw exception TypeError: Not enough arguments.
PASS observer.observe(null) threw exception TypeError: Not enough arguments.
PASS observer.observe(undefined) threw exception TypeError: Not enough arguments.
PASS observer.observe(document.body) threw exception TypeError: Not enough arguments.
PASS observer.observe(document.body, null) threw exception SyntaxError: An invalid or illegal string was specified..
PASS observer.observe(document.body, undefined) threw exception SyntaxError: An invalid or illegal string was specified..
PASS observer.observe(null, {attributes: true}) threw exception NotFoundError: An attempt was made to reference a Node in a context where it does not exist..
PASS observer.observe(undefined, {attributes: true}) threw exception NotFoundError: An attempt was made to reference a Node in a context where it does not exist..
PASS observer.observe(document.body, {subtree: true}) threw exception SyntaxError: An invalid or illegal string was specified..
PASS observer.observe(document.body, {childList: true, attributeOldValue: true}) threw exception SyntaxError: An invalid or illegal string was specified..
PASS observer.observe(document.body, {attributes: true, characterDataOldValue: true}) threw exception SyntaxError: An invalid or illegal string was specified..
PASS observer.observe(document.body, {characterData: true, attributeFilter: ["id"]}) threw exception SyntaxError: An invalid or illegal string was specified..
PASS successfullyParsed is true

TEST COMPLETE
